+++ bot/config.py
@@ -1,0 +1,5 @@
+# config.py
+
+# Your bot token from BotFather
+BOT_TOKEN = '7616802949:AAEVIoSBug1sJBosOzO3lJ13kVqN_82MKRI'
+CHAT_ID = "@suiTrending_boost"  # Replace with your actual chat ID
